requirements.txt
torch
numpy
matplotlib
torchvision
Pillow
omegaconf
tensorboardX